# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Backend developer test",
        "version": "1.0"
    },
    "paths": {
...
        "/api/books": {
            "get": {
                "tags": [
                    "BookController"
                ],
                "summary": "Books list",
                "operationId": "51d99736515804e100bc10bd753fe94f",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "result": {
                                        "summary": "An result object.",
                                        "value": {
                                            "-Mw895jMgCnMQvdc_sAD": {
                                                "name": "tutorial koding",
                                                "price": "89000"
                                            },
                                            "-Mw89UrB_YOq778EHzG6": {
                                                "name": "tutorial html",
                                                "price": "89000"
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "result": {
                                        "summary": "An result object.",
                                        "value": {
                                            "message": "Unauthorized."
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/books/{id}": {
            "get": {
                "tags": [
                    "BookController"
                ],
                "summary": "Books detail",
                "operationId": "627bfaca2a5011075332f5b4a267fbcf",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "Book id",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "result": {
                                        "summary": "An result object.",
                                        "value": {
                                            "name": "tutorial koding",
                                            "price": "89000"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Bad request",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "result": {
                                        "summary": "An result object.",
                                        "value": {
                                            "message": "Buku tidak ditemukan"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "result": {
                                        "summary": "An result object.",
                                        "value": {
                                            "message": "Unauthorized."
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/books/{id}/create": {
            "post": {
                "tags": [
                    "BookController"
                ],
                "summary": "Books add",
                "operationId": "9af78ec33456bf7030f7b896d9c21458",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "price": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "id": "Buku membaca",
                                    "name": "98000"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "result": {
                                        "summary": "An result object.",
                                        "value": {
                                            "-Mw895jMgCnMQvdc_sAD": {
                                                "name": "tutorial koding",
                                                "price": "89000"
                                            },
                                            "-Mw89UrB_YOq778EHzG6": {
                                                "name": "tutorial html",
                                                "price": "89000"
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Bad request",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "result": {
                                        "summary": "An result object.",
                                        "value": {
                                            "errors": {
                                                "name": [
                                                    "The name field is required."
                                                ],
                                                "price": [
                                                    "The price field is required."
                                                ]
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "result": {
                                        "summary": "An result object.",
                                        "value": {
                                            "message": "Unauthorized."
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/books/{id}/update": {
            "put": {
                "tags": [
                    "BookController"
                ],
                "summary": "Books update",
                "operationId": "e3d06edbf70b9cd0ee1c5abf692468e4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "Book id",
                        "required": true
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "price": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "id": "Buku membaca",
                                    "name": "98000"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "result": {
                                        "summary": "An result object.",
                                        "value": {
                                            "name": "Buku membaca",
                                            "price": "98000"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Bad request",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "result": {
                                        "summary": "An result object.",
                                        "value": {
                                            "errors": {
                                                "name": [
                                                    "The name field is required."
                                                ],
                                                "price": [
                                                    "The price field is required."
                                                ]
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Bad request",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "result": {
                                        "summary": "An result object.",
                                        "value": {
                                            "message": "Buku tidak ditemukan"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "result": {
                                        "summary": "An result object.",
                                        "value": {
                                            "message": "Unauthorized."
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/books/{id}/delete": {
            "delete": {
                "tags": [
                    "BookController"
                ],
                "summary": "Books delete",
                "operationId": "16d3b142d2e9ee91a702d96d28987272",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "Book id",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "result": {
                                        "summary": "An result object.",
                                        "value": {
                                            "name": "tutorial koding",
                                            "price": "89000"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Bad request",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "result": {
                                        "summary": "An result object.",
                                        "value": {
                                            "message": "Buku tidak ditemukan"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "result": {
                                        "summary": "An result object.",
                                        "value": {
                                            "message": "Unauthorized."
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
    }
}